Made som tests for Deck and refactored a bit
Removed handOutCards and replaced it with
a method that transfers n cards from a deck,
to a target deck.

Made tests for the majority of Decks functionality:
GetCardAtPosition
EnhancedForLoops
Checking number of cards in a deck
Creating a new deck
Creating a deck of program cards
Adding a collection of cards to a deck, etc.